Revert "b-teddy"
