--- .prettierrc
@@ -1,8 +1,0 @@
-{
-  "jsxSingleQuote": true,
-  "bracketSameLine": false,
-  "bracketSpacing": true,
-  "embeddedLanguageFormatting": "auto",
-  "semi": false,
-  "htmlWhitespaceSensitivity": "css"
-}
