--- index.js
@@ -1,1 +1,0 @@
-const oo = bar;
